+++ src/main/kotlin/com/github/sam0delkin/intellijpsa/psi/reference/PsaReferenceContributor.kt
@@ -55,7 +55,7 @@ class PsaReferenceContributor : PsiReferenceContributor() {
                         return emptyArray()
                     }
 
-                    if (!settings.supportsStaticCompletions || null === psaManager.getStaticCompletionConfigs()) {
+                    if (!settings.supportsStaticCompletions) {
                         return emptyArray()
                     }